fix more runtime warning

+++ pym2149/osc.py
@@ -118,7 +118,7 @@ class RationalDiff(BinDiff):
       return self.hold
     else:
       self.blockbuf.fill(0)
-      stepcount = fracfloor((self.block.framecount - 1 - stepindex) / stepsize) + 1
+      stepcount = fracfloor(np.int64(self.block.framecount - 1 - stepindex) / stepsize) + 1
       lcd = fraclcd(stepsize, stepindex)
       indices = -((-fracint(stepindex, lcd) - np.arange(stepcount) * fracint(stepsize, lcd)) // lcd)
       dc = self.ringcursor.currentdc()
